Tests

See merge request foxstudent100582/location-tracker-a!2

+++ app/src/main/java/ua/com/foxminded/locationtrackera/model/locations/UserLocation.java
@@ -11,9 +11,9 @@ public class UserLocation {
     public int locationId;
     public double latitude;
     public double longitude;
-    public double date;
+    public long date;
 
-    public UserLocation(double latitude, double longitude, double date) {
+    public UserLocation(double latitude, double longitude, long date) {
         this.latitude = latitude;
         this.longitude = longitude;
         this.date = date;
